Simplify further, and note how selecting top-levels is going awry.
Using href so we can select the base nav entries to travel to them. This
sucks, though, because it sacrifices the efficiencies of the single-page
app traversal. I'm taking measures (not yet quite working) so the hrefs are
only used in the base entries, not ones on the menus. We need to fix it
fully, instead. (See note about ContextNavigation.jsx in the next
paragraph.)

I'm realizing that the dropdown menu headers work without the hrefs when
they're embedded in a menu, rather than among the base navbar entries. This
suggests to me that there is something in ContextNavigation.jsx that is
inhibiting the `key` behavior (as also seems to be the case for active.item
border-color).