Return to including container as "header" in dropdown.
Using href doesn't work right. It gives the href to the entire menu, not
entry from which the menu comes.

So using the container-as-menu-header-entry is our awkward but usable
workaround for the problem getting the menu container entry itself to be
selectable.

The other remaining problem is related: highlighting to distinguish
"occupied" pages - pages on path of the currently visited page - is not
working. I believe this is because the cue being used is the `active` CSS
class, specifically `item.activated`, which is also used by the menu
entry selection mechanism.

I believe I can work around this by deriving another class, say "occupied",
and using that for the topmost entry distinction. That will suffice until I
can get the attention of someone who knows both React and Volto well enough
to solve this and the container selection problems properly.