Implement select-on-containing-entry, but stubborn.
Incorporate NavLink behavior in the Dropdown base entry using the `as=`
attribute. We can now click on the base entry in order to transit to that
page. Yay!

An unintended consequence is that the menus are automatically and
persistently open to present the currently visited page.

The problem may be connected with the `item.active` class being
automatically acquired by NavLinks that qualify for
`isInternalURL()` (found in @plone/volto/helpers), while `active` is also
used by `simple` Dropdownmenus for menus that are open.

It may instead be that the `as={NavLink}` disrupts the `simple`
Dropdownmenu closing behavior, but that wouldn't explain why the
menus (sometimes) open to the entry for the current page when navigating to
it using the URL directly, or the back or forward buttons, etc.

If nothing else, this implementation demonstrates an approach that mostly
works.